Add support for array of links with JSON HAL
I updated the test to more closely follow the example in the HAL specification:

http://tools.ietf.org/html/draft-kelly-json-hal-00#section-6

If the href of a link is an Array, then the representer returns an array of link objects per the HAL spec (http://tools.ietf.org/html/draft-kelly-json-hal-03#section-4.1.1):

```ruby
link :upsell do
  upsells # Upsells is an Array of Products
end

```

Note that each Product in the `upsells` array should specify how to represent itself. I added a new `as` option to the `property` method that allows a representer to specify how its value should be represented:

```ruby
ProductRepresenter = Module.new do
  include Roar::Representer::JSON::HAL
  property :id, :from => :href, :as => "/products/?"
  property :name
  property :title
end

```

The `as` option modifies the value of the property in the representation.

[Fixes #33]